+++ modules/optionals/datalake-buckets/tests/test_app.py
@@ -54,3 +54,25 @@ def test_invalid_retention_type(stack_defaults):
     os.environ["ADDF_PARAMETER_RETENTION_TYPE"] = "notvalid"
     with pytest.raises(Exception):
         import app  # noqa: F401
+
+
+def test_solution_description(stack_defaults):
+    os.environ["ADDF_PARAMETER_SOLUTION_ID"] = "SO123456"
+    os.environ["ADDF_PARAMETER_SOLUTION_NAME"] = "MY GREAT TEST"
+    os.environ["ADDF_PARAMETER_SOLUTION_VERSION"] = "v1.0.0"
+
+    import app
+
+    ver = app.generate_description()
+    assert ver == "(SO123456) MY GREAT TEST. Version v1.0.0"
+
+
+def test_solution_description_no_version(stack_defaults):
+    os.environ["ADDF_PARAMETER_SOLUTION_ID"] = "SO123456"
+    os.environ["ADDF_PARAMETER_SOLUTION_NAME"] = "MY GREAT TEST"
+    del os.environ["ADDF_PARAMETER_SOLUTION_VERSION"]
+
+    import app
+
+    ver = app.generate_description()
+    assert ver == "(SO123456) MY GREAT TEST"