+++ modules/post-processing/yolop-lane-detection/app.py
@@ -21,6 +21,20 @@ removal_policy = os.getenv(_param("REMOVAL_POLICY"), "")
 if not full_access_policy:
     raise ValueError("S3 Full Access Policy ARN is missing.")
 
+
+def generate_description() -> str:
+    soln_id = os.getenv("ADDF_PARAMETER_SOLUTION_ID", None)
+    soln_name = os.getenv("ADDF_PARAMETER_SOLUTION_NAME", None)
+    soln_version = os.getenv("ADDF_PARAMETER_SOLUTION_VERSION", None)
+
+    desc = "(SO9154) Autonomous Driving Data Framework (ADDF) - yolop-lane-det"
+    if soln_id and soln_name and soln_version:
+        desc = f"({soln_id}) {soln_name}. Version {soln_version}"
+    elif soln_id and soln_name:
+        desc = f"({soln_id}) {soln_name}"
+    return desc
+
+
 app = App()
 
 stack = LaneDetection(
@@ -34,6 +48,7 @@ stack = LaneDetection(
     ),
     s3_access_policy=full_access_policy,
     removal_policy=RemovalPolicy.RETAIN if removal_policy.upper() == "RETAIN" else RemovalPolicy.DESTROY,
+    stack_description=generate_description(),
 )