+++ modules/post-processing/yolop-lane-detection/tests/infra/test_app.py
@@ -31,3 +31,25 @@ def test_full_access_policy(stack_defaults):
         import app  # noqa: F401
 
         assert os.environ["ADDF_PARAMETER_FULL_ACCESS_POLICY_ARN"] == "arn:aws:policy:12345:XXX"
+
+
+def test_solution_description(stack_defaults):
+    os.environ["ADDF_PARAMETER_SOLUTION_ID"] = "SO123456"
+    os.environ["ADDF_PARAMETER_SOLUTION_NAME"] = "MY GREAT TEST"
+    os.environ["ADDF_PARAMETER_SOLUTION_VERSION"] = "v1.0.0"
+
+    import app
+
+    ver = app.generate_description()
+    assert ver == "(SO123456) MY GREAT TEST. Version v1.0.0"
+
+
+def test_solution_description_no_version(stack_defaults):
+    os.environ["ADDF_PARAMETER_SOLUTION_ID"] = "SO123456"
+    os.environ["ADDF_PARAMETER_SOLUTION_NAME"] = "MY GREAT TEST"
+    del os.environ["ADDF_PARAMETER_SOLUTION_VERSION"]
+
+    import app
+
+    ver = app.generate_description()
+    assert ver == "(SO123456) MY GREAT TEST"