+++ modules/sensor-extraction/ros-to-png/app.py
@@ -32,6 +32,20 @@ if resized_height:
 if not full_access_policy:
     raise ValueError("S3 Full Access Policy ARN is missing.")
 
+
+def generate_description() -> str:
+    soln_id = os.getenv("ADDF_PARAMETER_SOLUTION_ID", None)
+    soln_name = os.getenv("ADDF_PARAMETER_SOLUTION_NAME", None)
+    soln_version = os.getenv("ADDF_PARAMETER_SOLUTION_VERSION", None)
+
+    desc = "(SO9154) Autonomous Driving Data Framework (ADDF) - ros-to-png"
+    if soln_id and soln_name and soln_version:
+        desc = f"({soln_id}) {soln_name}. Version {soln_version}"
+    elif soln_id and soln_name:
+        desc = f"({soln_id}) {soln_name}"
+    return desc
+
+
 app = App()
 
 stack = RosToPngBatchJob(
@@ -51,7 +65,9 @@ stack = RosToPngBatchJob(
     removal_policy=RemovalPolicy.RETAIN if removal_policy.upper() == "RETAIN" else RemovalPolicy.DESTROY,
     resized_width=resized_width,  # type: ignore
     resized_height=resized_height,  # type: ignore
+    stack_description=generate_description(),
 )
+
 
 CfnOutput(
     scope=stack,